Renamed util folder to utils
