+++ pages/api/auth/[...nextAuth].ts
@@ -1,0 +1,12 @@
+import NextAuth from "next-auth/next";
+import EmailProvider from "next-auth/providers/email";
+
+export const options = {
+    providers: [
+        EmailProvider({
+            
+        })
+    ]
+}
+
+export default NextAuth(options);
